WIP: check update model issue
- Add request input
