+++ controllers/masthead.go
@@ -1,10 +1,13 @@
 package controllers
 
 import (
+	"fmt"
 	"net/http"
+	"reflect"
 	"time"
 
 	"example.com/m/models"
+	"example.com/m/requests"
 	"github.com/gin-gonic/gin"
 )
 
@@ -16,7 +19,7 @@ func GetMastheads(c *gin.Context) {
 }
 
 func CreateMasthead(c *gin.Context) {
-	var input models.CreateMastheadInput
+	var input requests.CreateMastheadInput
 
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -50,19 +53,48 @@ func GetMasthead(c *gin.Context) {
 func UpdateMasthead(c *gin.Context) {
 	var masthead models.Masthead
 
+	tmp := c.Request.FormValue("order")
+	fmt.Println(reflect.TypeOf(tmp))
+	if tmp == "" {
+		fmt.Print("hey!")
+	}
+
 	if err := models.DB.Where("id = ?", c.Param("id")).First(&masthead).Error; err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	var input models.UpdateMastheadInput
+	var input requests.UpdateMastheadInput
 
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	models.DB.Model(&masthead).Updates(input)
+	// models.DB.Model(&masthead).Updates(map[string]interface{}{
+	// 	"ImageURL": input.ImageURL,
+	// 	"Link":     input.Link,
+	// 	"Order":    input.Order,
+	// 	"Status":   input.Status,
+	// })
+
+	fmt.Println(input)
+
+	var m = map[string]interface{}{
+		"ImageURL": input.ImageURL,
+		"Link":     input.Link,
+		"Order":    input.Order,
+		"Status":   input.Status,
+	}
+
+	for k, v := range m {
+		if v == 0 {
+			delete(m, k)
+		}
+	}
+
+	fmt.Println(m)
+	models.DB.Model(&masthead).Updates(m)
 
 	c.JSON(http.StatusOK, gin.H{"data": masthead})
 }
